requirements.txt
streamlit
streamlit-chat
python-dotenv
faiss-cpu
google-generativeai

